Crate lignin_html

An HTML renderer for lignin that does some syntactic and no semantic validation.

Escaping is performed automatically where necessary, but the output isn't guaranteed to be minimal.

lignin-html is not round-trip-safe regarding any HTML parser implementation.
This is impossible for any HTML renderer that accepts adjacent Node::Texts, but maybe should still be noted explicitly.

About the Documentation

HTML terms are bold italic and linked to the HTML specification (as of 2021-03-04).

Caveats

In HTML comments, if illegal comment text is encountered, certain dashes (-) are silently replaced with equal signs (=) and pipe characters (|) are silently inserted around the comment text as necessary. See Comments.

Originally I was going to use zero width non-joiner and zero width joiner characters for this, to make the comment resemble the original better, but this could be a very bad idea if any transport in-between strips Unicode.

Re-exports

pub use lignin;

Structs

Error

Functions

render_document
Renders vdom into target as HTML document with DOCTYPE*.

render_fragment
Renders vdom into target as HTML fragment without DOCTYPE*.