Function lignin_html::render_document

pub fn render_document<'a, S: ThreadSafety>(
    vdom: &'a Node<'a, S>, 
    target: &mut impl Write, 
    depth_limit: usize
) -> Result<(), Error<'a, S>>

Renders vdom into target as HTML document with DOCTYPE*.

depth_limit is measured in Nodes and must be at least 1 to not error on it.

Caveats

See render_fragment#caveats.

Errors

Iff vdom is found to represent invalid HTML.

Warning: This function succeeding does not guarantee that the produced HTML is fully valid!